Refactor transfer/image command buffer commands
